pyproject.toml
[tool.poetry]
name = "bugzoo"
version = "2.3.0"
description = "A platform for safely studying historical software versions."
authors = ["Chris Timperley <ctimperley@cmu.edu>"]
license = "MIT"
repository = "https://github.com/squaresLab/BugZoo"
readme = "README.md"
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.scripts]
bugzoo = 'bugzoo.cli.app:main'
bugzood = 'bugzoo.server:main'

[tool.poetry.dependencies]
tabulate = ">=0.9.0"
pyyaml = ">=6.0.1"
python = ">=3.11,<4"
attrs = ">=19.3.0"
cement = "^3.0"
psutil = ">=5.0.0"
chardet = ">=3.0.4"
docker = ">=3.1.4"
dockerpty = ">=0.4.1"
GitPython = ">=3.1.23"
flask = "^2.2.5"
deprecated = "^1.2.6"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
flake8-bugbear = "^23.9.16"
flake8-comprehensions = "^3.14.0"
flake8-future-annotations = "^1.1.0"
flake8-import-order = "^0.18.2"
flake8-quotes = "^3.3.2"
icecream = "^2.1.3"
mypy = ">=1.10.1"
nox = "^2023.4.22"
pep8-naming = "^0.13.3"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
python = "^3.11"
ruff = "^0.1.6"
types-pyyaml = "*"

[tool.mypy]
ignore_missing_imports = false
strict = true

[[tool.mypy.overrides]]
module = "bugzoo.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "cement.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pyroglyph.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "comby.*"
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["src"]
select = [
    "ALL",  # includes all rules (including new ones)
]
per-file-ignores = {"*/__init__.py" = ["F401"]}
extend-ignore = [
    "ANN101",   # missing type annotation for 'self' in method
    "ANN102",   # missing type annotation for 'cls' in classmethod
    "B024",     # use of abstract base class without abstract methods
    "D100",     # missing docstring in public module
    "D101",     # missing docstring in public class
    "D102",     # missing docstring in public method
    "D103",     # missing docstring in public function
    "D104",     # missing docstring in public package
    "D105",     # missing docstring in magic method
    "D203",     # one-blank-line-before-class
    "D213",     # multi-line-summary-second-line
    "D401",     # first line of docstring should be in imperative mood
    "D413",     # missing blank line after last section
    "FIX001",   # unresolved FIXME
    "FIX002",   # TODO found
    "TD001",    # invalid TODO tag: `FIXME`
    "TD002",    # missing author in TODO
    "TD003",    # missing issue link on line following TODO
    "TD004",    # missing colon in TODO
    "TRY003",   # avoid specifying long messages outside the exception class
    "S101",     # use of assertions
    "SLF001",   # private member accessed
    "T201",     # use of print
    "D204",     # 1 blank line required after class docstring
    "C901",     # function is too complex (cyclomatic complexity)
    "PLR0912",  # too many branches
    "PGH003",   # use specific rule codes when ignoring type issues
    "FBT001",   # boolean typed positional argument in function definition
    "ARG001",   # unused function argument
    "PLR0913",  # too many arguments in function definition
    "PLR0911",  # too many return statements
    "PYI041",   # use `float` instead of `int | float`
    "D106",     # missing docstring in public nested class
    "S311",     # standard pseudo-random generators are not suitable for security/cryptographic purposes
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
